fixed sentiment selection edge case
